labas rytas

+++ bankas_l4/routes/web.php
@@ -19,27 +19,25 @@ Route::get('/', function () {
     return view('welcome');
 });
 
+Route::get('/home', [C::class, 'index'])->name('clients-index');
+
 
 // ========================== CLIENTS ==========================
 
-Route::get('/clients', [C::class, 'index'])->name('clients-index');
-Route::get('/clients/create', [C::class, 'create'])->name('clients-create');
-Route::post('/clients', [C::class, 'store'])->name('clients-store');
-Route::get('/clients/edit/{client}', [C::class, 'edit'])->name('clients-edit');
-Route::put('/clients/{client}', [C::class, 'update'])->name('clients-update');
-Route::delete('/clients/{client}', [C::class, 'destroy'])->name('clients-delete');
-Route::get('/clients/show/{id}', [C::class, 'show'])->name('clients-show');
+Route::get('/clients', [C::class, 'index'])->name('clients-index')->middleware('rp:user');
+Route::get('/clients/create', [C::class, 'create'])->name('clients-create')->middleware('rp:user');
+Route::post('/clients', [C::class, 'store'])->name('clients-store')->middleware('rp:user');
+Route::get('/clients/edit/{client}', [C::class, 'edit'])->name('clients-edit')->middleware('rp:admin');
+Route::put('/clients/{client}', [C::class, 'update'])->name('clients-update')->middleware('rp:admin');
+Route::delete('/clients/{client}', [C::class, 'destroy'])->name('clients-delete')->middleware('rp:admin');
+Route::get('/clients/show/{id}', [C::class, 'show'])->name('clients-show')->middleware('rp:user');
 
 // ========================== FUNDS ==========================
 
 Route::get('/clients/funds/{client}', [C::class, 'funds'])->name('clients-funds');
 Route::put('/clients/addfunds/{client}', [C::class, 'addfunds'])->name('clients-addfunds');
 Route::put('/clients/withdrawfunds/{client}', [C::class, 'withdrawfunds'])->name('clients-withdrawfunds');
-
-
 Route::put('/clients/editfunds/{client}', [C::class, 'editfunds'])->name('clients-editfunds');
 
+// ========================== VITE ==========================
 Auth::routes();
-
-// Route::get('/home', [App\Http\Controllers\HomeController::class, 'index'])->name('home');
-Route::get('/home', [C::class, 'index'])->name('clients-index');
